Add projeto example-fixwo-easybpms

+++ example-fixwo-easybpms/src/main/java/org/fixwo/services/AssociarCliente.java
@@ -2,7 +2,7 @@ package org.fixwo.services;
 
 public class AssociarCliente {
 	
-	public String run(Integer idOcorrencia){
+	public String run(Long idOcorrencia){
 		/*Receber o id da ocorrencia para conseguir buscar a area cadastrada para ela.
 		  Buscar na entidade area o cliente responsavel por ela. Setar na entidade
 		  ocorrencia, no atributo tenancy o id do cliente. Retornar o tenancy*/
