+++ example-fixwo-easybpms/src/main/java/org/fixwo/services/BuscarArea.java
@@ -1,7 +1,7 @@
 package org.fixwo.services;
 
 public class BuscarArea {
-	public boolean run(Integer idOcorrencia){
+	public boolean run(Long idOcorrencia){
 		
 	/*Receber o id da ocorrencia para conseguir buscar a geolocalizacao armazenada
 	para a ocorrencia. Fazer uma busca das areas cadastradas com aquelas coordenadas.
